chore(engine): avoid NPE while reconfiguring the cleanup job
related to CAM-11147

+++ engine/src/main/java/org/camunda/bpm/engine/impl/persistence/entity/EverLivingJobEntity.java
@@ -75,8 +75,14 @@ public class EverLivingJobEntity extends JobEntity {
     }
 
     if (exceptionByteArrayIdToDelete != null) {
-      ByteArrayEntity byteArray = commandContext.getDbEntityManager().selectById(ByteArrayEntity.class, exceptionByteArrayIdToDelete);
-      commandContext.getDbEntityManager().delete(byteArray);
+      ByteArrayEntity byteArray = commandContext.getDbEntityManager()
+          .selectById(ByteArrayEntity.class, exceptionByteArrayIdToDelete);
+
+      // Avoid NPE when the job was reconfigured by another
+      // node in the meantime
+      if (byteArray != null) {
+        commandContext.getDbEntityManager().delete(byteArray);
+      }
     }
   }
